requirements.txt
pdfminer.six
pdfplumber
python-pptx
python-docx
pytesseract
pdf2image
Pillow
